SupportArticleDialog: async load the component on first display
